python/pyproject.toml
[build-system]
requires = ["setuptools>=61.0", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "ls-project"
version = "0.1.0"
description = ""
readme = "README.md"
requires-python = ">=3.11, <=3.13"
license = {text = "MIT"}
authors = [
    {name = "Your Name", email = "your.email@example.com"},
]
classifiers = [
    "Development Status :: 3 - Alpha",
    "Intended Audience :: Developers",
    "License :: OSI Approved :: MIT License",
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3.8",
    "Programming Language :: Python :: 3.9",
    "Programming Language :: Python :: 3.10",
    "Programming Language :: Python :: 3.11",
    "Programming Language :: Python :: 3.12",
]
dependencies = [
    "openai-agents>=0.6.0",
    "braintrust>=0.4.0",
    "claude-agent-sdk>=0.1.19",
    "deepagents>=0.3.3",
    "jupyter>=1.1.1",
    "langchain>=1.2.0",
    "langchain-community>=0.4.1",
    "langchain-openai>=0.3.0",
    "langgraph-sdk>=0.3.1",
    "langsmith>=0.6.0",
    "load-dotenv>=0.1.0",
    "openai>=2.14.0",
    "langsmith-fetch>=0.3.1",
    "numpy>=1.24.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=7.0",
    "black>=23.0",
    "ruff>=0.1.0",
]

[tool.setuptools]
packages = []

[tool.black]
line-length = 88
target-version = ['py38', 'py39', 'py310', 'py311']

[tool.ruff]
line-length = 88
target-version = "py38"

[tool.ruff.lint]
select = ["E", "F", "I", "N", "W"]
ignore = []

[tool.pytest.ini_options]
testpaths = ["tests"]
python_files = ["test_*.py", "*_test.py"]

